requirements.txt
# Uses Python 3.9.5

# Base libraries
# TODO: Add libraries that are used in the project

# Code style
black==22.3
flake8
isort==5.10.1
pytest
pylint==2.13
mypy==0.782
pre-commit==2.20.0

# Pytorch
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.0
torchmetrics==0.10.0

# Data manipulation
pandas==1.5
scikit-learn==1.1.2

# Experiment tracking and reproducibility
wandb==0.12.5
protobuf==3.20

# Data visualization
matplotlib==3.5.1
seaborn==0.11.2

# argparse
argparse==1.4.0

# PIL
Pillow==8.4.0
imageio==2.10.4
tqdm==4.62.3